refactor: extract LoginMixin and login detection types
Move 18 login/auth methods, LoginDetectionReason enum,
LoginDetectionResult dataclass, and selector constants from
__init__.py into a new LoginMixin in utils/login_mixin.py.

Part of breaking down __init__.py to reduce monolith size.
